+++ slim/base/app.py
@@ -55,12 +55,12 @@ class CORSOptions:
 
 class Application:
     def __init__(self, *, cookies_secret: bytes, log_level=logging.DEBUG, session_cls=CookieSession,
-                 permission: Optional['Permissions'], client_max_size=2 * 1024 * 1024,
+                 permission: Optional['Permissions'] = None, client_max_size=2 * 1024 * 1024,
                  cors_options: Optional[Union[CORSOptions, List[CORSOptions]]] = None):
         """
         :param cookies_secret:
         :param log_level:
-        :param permission: `ALL_PERMISSION`, `EMPTY_PERMISSION` or a `Permissions` object
+        :param permission: `ALL_PERMISSION`, `NO_PERMISSION` or a `Permissions` object
         :param session_cls:
         :param client_max_size: 2MB is default client_max_body_size of nginx
         """
@@ -71,7 +71,7 @@ class Application:
         if permission is ALL_PERMISSION:
             logger.warning('app.permission is ALL_PERMISSION, it means everyone has all permissions for any table')
             self.permission = Permissions(self)
-            self.permission.add(Ability(None, {'*': '*'}))  # everyone has all permission for all table
+            self.permission.add(None, Ability({'*': '*'}))  # everyone has all permission for all table
         elif permission is None or permission is NO_PERMISSION:
             self.permission = Permissions(self)  # empty
         else: