+++ slim/utils/__init__.py
@@ -7,7 +7,7 @@ import time
 from typing import Optional
 
 from .async_run import *
-from .binhex import to_bin, to_hex
+from .binhex import to_bin, to_hex, get_bytes_from_blob
 from .cls_init import MetaClassForInit
 from .pagination import pagination_calc
 from .state_obj import StateObject
@@ -42,18 +42,6 @@ def dict_filter_inplace(obj, keys):
 
     for i in to_remove:
         del obj[i]
-
-
-def get_bytes_from_blob(val) -> Optional[bytes]:
-    """ 不同数据库从blob拿出的数据有所差别，有的是memoryview有的是bytes """
-    if isinstance(val, bytes):
-        return val
-    elif isinstance(val, memoryview):
-        return val.tobytes()
-    elif val is None:
-        return None
-    else:
-        raise TypeError('invalid type for get bytes')
 
 
 def time_readable():