+++ slim_cli/template/permissions/role_define.py
@@ -1,0 +1,7 @@
+from slim.utils import StateObject
+
+
+class ACCESS_ROLE(StateObject):
+    VISITOR = None
+    NORMAL_USER = 'normal_user'
+